#5514 add null check for confirm email

+++ src/main/java/edu/harvard/iq/dataverse/engine/command/impl/MergeInAccountCommand.java
@@ -149,8 +149,11 @@ public class MergeInAccountCommand extends AbstractVoidCommand {
         
         //ConfirmEmailData  
         
-        ConfirmEmailData confirmEmailData = ctxt.confirmEmail().findSingleConfirmEmailDataByUser(consumedAU);       
-        ctxt.em().remove(confirmEmailData);
+        ConfirmEmailData confirmEmailData = ctxt.confirmEmail().findSingleConfirmEmailDataByUser(consumedAU); 
+        if (confirmEmailData != null){
+            ctxt.em().remove(confirmEmailData);
+        }
+
         
         //Access Request is not an entity. have to update with native query
